fix: handle HubSpot 429 rate limits without dropping jobs
Separate 429 and 5xx retry logic in _retry_loop. 429s now retry up to
8 times honoring Retry-After (clamped 10-60s). On exhaustion, raises
HubSpotRateLimitExhausted instead of dying permanently. All public
webhook handlers catch this exception and re-enqueue themselves on the
long queue. Moves all webhook-triggered jobs from short to long queue
to reduce concurrent API call pressure.

+++ ivm/integrations/hubspot/webhook.py
@@ -162,7 +162,7 @@ def _route_event(event: dict[str, Any]) -> None:
     kwargs["hubspot_user_id"] = user_id
 
     try:
-        frappe.enqueue(method, queue="short", **kwargs)
+        frappe.enqueue(method, queue="long", **kwargs)
     except Exception:
         frappe.log_error(
             title=f"HubSpot webhook: failed to enqueue {subscription_type} "
